Created src/models, refactoring
